requirements.txt
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
py-bcrypt==0.4
SQLAlchemy==1.2.4
Werkzeug==0.14.1
WTForms==2.1
